+++ src/hopfield/update/asynchronous.py
@@ -22,7 +22,7 @@ def rand_indexes(N: int,
 def asynch(state_orig: NDArray[np.float64], 
            W: NDArray[np.float64],
            indexes: NDArray[np.int64]
-           ) ->tuple[NDArray[np.int8], NDArray[np.float64]]:
+           ) ->tuple[NDArray[np.int64], NDArray[np.float64]]:
     """
     asynchronous update; calulating the local field of neurons one by one 
     randomly until given number of sweeps or convergence.
@@ -57,13 +57,13 @@ def asynch(state_orig: NDArray[np.float64],
         if not changed:
             break
 
-    return state.astype(np.int8), energy_arr[:count]
+    return state.astype(np.int64), energy_arr[:count]
 
 
-def update_asynch(state_orig: NDArray[np.int8], 
+def update_asynch(state_orig: NDArray[np.int64], 
                   W: NDArray[np.float64],
                   rng: np.random.Generator,
-                  ) ->tuple[NDArray[np.int8], NDArray[np.float64]]:
+                  ) ->tuple[NDArray[np.int64], NDArray[np.float64]]:
 
     sweeps = 1000
     indexes_arr = rand_indexes(state_orig.size, sweeps, rng)
